refactor : 도매인 개념으로 폴더 분리
