+++ src/main/java/app/bottlenote/follow/repository/follow/FollowRepository.java
@@ -1,4 +1,4 @@
-package app.bottlenote.follow.repository;
+package app.bottlenote.follow.repository.follow;
 
 import app.bottlenote.follow.domain.Follow;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface FollowCommandRepository extends JpaRepository<Follow, Long>, CustomFollowRepository {
+public interface FollowRepository extends JpaRepository<Follow, Long>, CustomFollowRepository {
 	@Query("SELECT f FROM Follow f JOIN FETCH f.user JOIN FETCH f.followUser WHERE f.user.id = :userId AND f.followUser.id = :followUserId")
 	Optional<Follow> findByUserIdAndFollowUserIdWithFetch(Long userId, Long followUserId);
 }
