+++ src/main/java/app/bottlenote/follow/service/FollowerService.java
@@ -1,0 +1,28 @@
+package app.bottlenote.follow.service;
+
+import app.bottlenote.follow.dto.request.FollowPageableRequest;
+import app.bottlenote.follow.dto.response.FollowSearchResponse;
+import app.bottlenote.follow.repository.follower.FollowerRepository;
+import app.bottlenote.global.service.cursor.PageResponse;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Slf4j
+@RequiredArgsConstructor
+@Service
+public class FollowerService {
+
+
+	private final FollowerRepository followerRepository;
+
+	@Transactional(readOnly = true)
+	public PageResponse<FollowSearchResponse> findFollowerList(Long userId, FollowPageableRequest pageableRequest) {
+
+		PageResponse<FollowSearchResponse> followerList = followerRepository.findFollowerList(userId, pageableRequest);
+
+		return followerList;
+	}
+
+}
